build: set the description of all projects
Previously, the description was inherited from camunda-release-parent.
Since that description was not accurate, we overwrote this to be empty
instead. This is not valid and can't be published on maven central, as
per the requirements:  https://central.sonatype.org/publish/requirements/#project-name-description-and-url

Instead of coming up with a sensible description for each project, here
we just set the description to be the name of the artifact.