ci: disable checkstyle for java client
Checkstyle has dropped support for pre-11 JDK versions. Since the
clients/java module is build on JDK 8, we need to disable the checkstyle
plugin altogether. It was already skipped, but the skip logic is
implemented by the plugin itself. And so this no longer works either.

Alternatively, we could've changed the skip logic for the checkstyle
plugin by replacing this logic with additional profiles. However, this
requires more thorough changes (e.g. removing the autoFormat profile
because it is active by default and would be disabled when another
profile is activated). These changes would also lead to more
duplication and more complex skip logic then we've defined at this time.

Disabling the plugin using a profile in the clients/java module only,
was the simplest solution.