test: write to log from `StreamProcessorRule` executed on actor thread
Since writing records sets timestamps and we want these timestamps to
be controlled by the actor clock of the `StreamProcessorRule`, writing
itself must happen within an actor thread.

Here we submit a helper actor from the `StreamProcessingComposite`
which is only used to execute the `write`s on an actor thread, while
building the record and `LogStreamRecordWriter` happens outside the
actor.