test(engine): support decision tables without IO names
A decision tables' inputs and outputs must have an id and can have a
name. The name (or label) is not mandatory.

This adds a failing test case that deploys a version of the jedi_or_sith
decision where all inputis and outputs no longer have names or labels.

The evaluated decision result should still be written to the log as
normal, but with an empty string as name.