feat(gateway-protocol): add new deploy resource rpc
The deploy process rpc was able to deploy non-process resources (like
dmn) simply by passing in a different resource. Even though that works,
the related metadata cannot be responded, because it is process
specific (i.e. decision metadata cannot be provided). A new deploy
resource rpc is therefore necessary.

The new deploy resource rpc is designed to be expandable to provide the
metadata of new types in the future. This could be achieved using the
oneof keyword.