fix(dmn): favor label as output name
The outputs of decision tables can have a name and can have a label,
according to the following rules:
- label is optional
- name is optional, unless the table has multiple outputs

The Modeler and Operate only need 1 name to display, but somehow we
should decide which one: the label or the name.

The Modeler favors the label if it is available it will be used, even if
a name is specified.

Zeebe should use this same behavior to decide which name to store in the
EvaluatedOutput of a DecisionEvaluation. This way Operate can directly
use the 'correct' name without having to parse the XML of the DMN model
itself.