refactor(dmn): extract evalute decision method
This method needs to be used in the upcoming decision table tests. These
tests will verify that a decision table with inputs that are labeled or
not have a different name in the evaluted input. (and likewise will test
similar behavior for outputs). These tests will need to be able to
evaluate decisions and it makes sense to just reuse this logic.