feat(protocol): add immutable protocol generation
Adds immutable protocol generation to the Zeebe record protocol. This is
done via an annotation process, Immutables, which requires two
annotations on each leaf types:

- `@Value.Immutable`
- `@ImmutableProtocol`

All leaf types - that is, types expected to have a single, direct, concrete
implementations - need to be annotated with both values to ensure that
there is always one immutable variant available, especially for
deserialization.

Meta-types such as `ProcessInstanceRelated`, for example, do not need to
be annotated but can be left as is. This may of course change in the
future.

This commit is the first step to removing the immutable protocol from
protocol-jackson, but for now both are existing at the same time.