fix(dmn): use expression as input name when label undefined
The inputs of decision tables can have an input, but these are not
mandatory. If it's missing, the Modeler shows the expression (if filled
in). Operate would like to have the same behavior as the Modeler in
showcasing it.

Zeebe can help Operate by deciding how to store the input name. This way
Opeate can directly use the 'correct' name without having to parse the
XML of the DMN model itself. This also makes the recordstream readable
in a similar fashion.