feat(protocol-util): adds new protocol-util module
Adds a new module for protocol utilities, which for now provides a
centralized way to discover type mappings, e.g. mapping value type to
values or intents, or abstract protocol types to concrete protocol
types.

These will be used in `protocol-jackson` in the next step, and later on
in other modules, e.g. to generate random records.