feat(gateway): support deploy resource rpc
A new rpc can be independently implemented by the clients, but that's
not true for the gateway. It should immediately support the new rpc.

Implementation was pretty straightfoward, with lots of code duplicated
from the existing DeployProcess rpc. The BrokerDeployProcessRequest can
simply be renamed as the broker's api doesn't change.

The only real difference happens in the ResponseMapper, which now also
has to map decisions and required decision into the new response object.