refactor(protocol-jackson): use new immutable protocol
Removes the old immutable `protocol-jackson`, and use the new one found
directly in the `protocol` module.

This completely changes how the module is used. The API is now to create
a new `ObjectMapper` and register an instance of `ZeebeProtocolModule`,
which knows how to serialize and deserialize records using the new
immutable protocol. This is achieved by reusing the mapping utilities
found in `protocol-util`.