refactor: propagate `UnrecoverableException`s during processing
With this change, unrecoverable exceptions that are thrown while
processing a command are directly thrown again and end up in
the `StreamProcessor`, where we already have logic to initiate a
transition to inactive.

Previously, unrecoverable exceptions were handled like all other
exceptions, leading to a retry and then an attempt to blacklist the
instance.