docs(gateway-protocol): clarify Deploy Resource behavior
Similar to the deprecated DeployProcess rpc, the documentation on the
DeployResource rpc should be clear about what users can expect.

Since the deploy logic is the same between the rpcs, we can continue to
use the same description of behavior (atomic deployment).

Note that we currently have a bug that undoes this behavior. However,
bugs are exceptions on our expected behavior. The documentation should
describe the expected behavior, whether we can currently adhere to it or
not.

The possible error scenarios stay the same, although the description
needed an update to generalize over different types of resources
(processes, decision models, etc).