feat: align deploy command return type naming
This replaces the return types of the newDeployCommand client method to
no longer be process-specific. This is important because the
newDeployCommand allows you to deploy any resource (process, decision,
etc).

The change is non-breaking and therefore non-intrusive for existing
users. However, the existing interfaces that refer to process have been
deprecated for removal. Users are instructed how to resolve the
deprecation through javadoc.