feat: use deploy resource rpc in deploy command
Switch the implementation of the java client's Deploy Command, from
calling the DeployProcess RPC to calling the DeployResource RPC.

This can be considered a refactoring, because nothing changes for the
user. It swaps out one implementation for another. Nevertheless, I've
marked this as feat, to highlight that this not just a simple
refactoring.

The DeployProcessTest has been updated as well, because it tests for the
implementation instead of the behavior. It verifies that the gateway
receives a specific request with expected data, instead of checking that
the response provided by the client contains specific data. This should
now test that DeployResource request is received by the gateway. Since
this test actually tests sending of the DeployResource request, I've
also renamed it to DeployResourceTest and made it clear in the tests
that we don't really care about processes, but rather about deploying
resources.