test: verify process responses for deploy command
Tests should generally only fail for 1 reason, but the first test case
(`shouldDeployResourceFromFile`) verifies multiple unrelated things.

To align with the other test cases in this class, it makes sense that
this test case only verifies that the gateway service was called with a
specific request.

We can extract the verification of the response into a separate test.
This can also be applied to the shouldDeployMultipleResources test case.