feat(clients/java): support deployed decision/drg metadata
The DeployResource command can be used to deploy processes and decision
models. The response contains metadata about the deployment, that is
process metadata in case one or processes were deployed, and/or decision
and decision requirements metadata in case one or more decision models
were deployed. Users should be able to extract this decision and
decision requirements metadata when they deploy dmn resources.

This commit introduces decisions and decisionRequirements to the
DeploymentEvent (which is the result of sending a DeployCommand for our
users). The Decision and DecisionRequirements interfaces are implemented
in exactly the same way as the Process interface. Also note that the
equals/hashcode/toString methods are generated to match the ProcessImpl.

New tests are written to show case the result of deploying a DMN
decision model (in this case it results in 1 decision requirements graph
and 2 decisions).