test(qa): deploy dmn using java client
This test is an acceptance test that verifies that the java client can
deploy a dmn decision model using the newDeployCommand client method.

It verifies that the model was resource was parsed and deployed,
resulting in a response that contains metadata of the deployed decision
requirements graph and the decisions it contains.