fix: don't notify listeners about replications of old role
Whenever a raft partition transitions to a new role, we must reset
`missedSnapshotReplicationEvents` so that registering new snapshot
replication listeners does not trigger the listener for snapshot
replication events that occurred for a different role.

This is to guard against a known case where raft received a snapshot and
transitioned to leader before role change and snapshot replication
listeners were registered. Once the registration of the snapshot
replication listener registration went through, the zeebe partition
transitioned to follower, while the raft partition was leader.