fix: only trigger snapshot replication listener on register if follower
Because we keep track of missed snapshot replication events to notify
listeners on registering, the partition is not necessarily in follower
role at that point. This breaks the assumption of the listener, so we
add a condition here to only trigger listeners on register if the raft
partition is still follower.