fix(clients/java): introduce newDeployResourceCommand
We discovered a compatibility issue between a newer java client and an
older gateway. If the gateway does not yet support the new
DeployResource RPC, then the newer client shouldn't call it, but instead
call the legacy DeployProcess RPC.

There's no way to safeguard against this without either the cost of
maintaining legacy code indefinitely, or asking users to actively switch
to the new RPC.

This introduces a newDeployResourceCommand method to ZeebeClient that
users should use to interact with the new DeployResource RPC.

It also deprecates the existing newDeployCommand to inform users that
they should switch. We will remove this method in the future.