feat(clients/java): optimize deployment metadata extraction
Maps the metadata in the deployment response more optimally, by only
iterating over the elements once.

This makes use of the `oneof` keyword in the Metadata message
definition, making it a mutually exclusive and thus offers a switch case
option.

Note that the type of metadata might not known. In that case, I think it
simply leads to a NullPointerException (at least that's what the
generated code looks like to me). But, in the off change that this
results in the METADATA_NOT_SET or default switch case branches, we can
use the opportunity to warn the user that they should probably
update their zeebe-client-java dependency.