refactor(protocol): map immutable protocol via annotations
Maps the immutable protocol directly via annotations. Instead of relying
on ClassGraph to build a map of all the types, we can simply map the
abstract, concrete, and builder types directly via annotations. These
can be used at runtime for deserialization, random generation, etc.

Here we are then adding a little bit of overhead when creating new
values, but hopefully reducing the amount of maintenance, as well as
making things a little bit more explicit.