deps: explicitly depend on google-java-format
Dependabot does not control the version of the google-java-format plugin
used by spotless. By adding an explicit maven dependency on the plugin
we can make Dependabot control it.

Note that adding this dependency doesn't change anything to the rest of
the project. It also doesn't really interfere with this specific plugin,
as it would in the worst case, simply add the same classes twice on the
classpath. By using a property, we can control that the depended version
is the same as the one loaded inside spotless.