feat(engine): allow appending/prepending evaluation contexts
This commit allows chaining instances of EvaluationContext together.
It also allows to prepend or append an additional EvaluationContext to an
existing ExpressionProcessor. This is in preparation to further developments
which will make use of this feature to combine variables stored in the variable
state and variables stored in the event trigger state, such that both are accessible
during evaluation.

To make this possible, a new interface ContextLookup was introduced. The existing class
VariableStateEvaluationContext was refactored to implement ContextLookup instead of EvaluationContext.

Also, two new methods and a new private constructor were added to ExpressionProcessor.
Together they allow composing complex ContextLookup lambdas. One thing that is not very intuitive
is that the context lookup is used to combine the evaluation contexts.
This was the best compromise of keeping the change small, while improving the semantic of the involved
classes and methods.

These changes shall be backported, which is the motivation for keeping them small.