build(clients/go): append v8 major version to module name
Go enforces a rather strange and stupid naming convention for major
version releases. It requires the module name to contain the major
version (if it isn't v0 or v1). Since we just moved to v8, we need to
make it part of the module name.

This commit changes the module name, updates the internal imports,
and updates the build and test scripts, accordingly.

See https://donatstudios.com/Go-v2-Modules for more info.