fix(engine): add grace period to detect and of processing
This commit adds a grace period after the end of processing has been reached.
Only when no new records are added to the logstream during the grace period, do
we proceed with the test.

This mechanism is a bit crude, but it has been utilized to great effect in
zeebe-process-test already. There we tried out different implementations and
found that this mechanism works best.

Downside is that this will make the tests run longer.
Also the exact value for GRACE_PERIOD may need to be tuned. The 50ms were
also taken from experiences in zeebe-process-test. But there we are working
with an in memory engine, so the value might not be optimal here.

(cherry picked from commit 0e61a9eb47cfaa7938e7fbdbeceb4b7005dddefe)