build: workaround plexus-util:1.1 injection
Ensure we always resolve plexus-util:1.1 when going offline. This is due
to Maven 3 injecting this artifact into all plugins and extensions which
do not have an explicit dependency on plexus-util. Since this is a
dynamic dependency (i.e. it's not declared explicitly in any POM), the
go-offline goal cannot resolve it and doesn't know it has to download
it. To workaround it we add an explicit dependency, with a note to never
update it.

The configuration to ignore the unused dynamic dependencies in the
prepare-offline profile is also moved to the profile instead of in the
shared configuration block.