fix(broker): do not log transition failure due to term mismatch as error
If a transition failed due to a term mismatch, there will be a new transition following. There is no need to log it as an error because conceptually it is not an error. It is expected to fail to prevent inconsistencies. When this failure happens, it is ignored and ZeebePartition continue with the next transition.

(cherry picked from commit 713d2ee29c8e1ee4cce24086ac736ab7063b1fea)