fix(dmn): prevent duplicate key insertion for dmn
To make sure we keep our data consistent we should make sure we don't store duplicate values into the state. The DMN resources were missing the required checks to prevent this. We would always try to insert the resources, disregarding if it is a duplicate. This change filters out the duplicate records and guarantees we only store the non-duplicates.

(cherry picked from commit 96810c816c94755324b44e0055dc3c2315dc4687)