test(test-util): reset timer to default on reset call
When reseting the exporting, reset the maximum wait time to default as
well as the records. This will ensure that things like the
`RecordingExporterTestWatcher`, which does call reset after each test,
will also reset the timeout. This may fix some flaky tests where we
sometimes change or reduce the timeout, which is then global for that
forked process, and since ordering is random it's very hard to
pinpoint/reproduce.