build(bom): support mvn spotless:apply
The 'spotless' prefix was not available, because the first project that
is run, is not the root-project but the bom. And the BOM did not have
knowledge of the spotless plugin in its build.

Adding the plugin like this to the build doesn't do anything else except
register the plugin for usage. The executions are still controlled by
the profiles defined in the parent-pom.