security: ignore netty-handler SNI
This also ignores netty-handler's disabled SNI by default. This is not
of particular concern to us because SNI is in fact enabled. And even
when disabled, we only serve a single virtual host per port - i.e. we
only server one certificate per port - so there would be no possible
confusion about serving the wrong certificate.