build(bom): fix spotless config
79cf9e0292a0660ada43856b88e55aedd8ec751b had broken the bom's pom.xml
because it now specified spotless without specifying a version. At the
time, I thought that was correct, because I expected the bom to inherit
from the parent pom. However, the parent of the bom is not
parent/pom.xml. Instead, the parent of parent/pom.xml is bom/pom.xml.

To fix the spotless definition in bom/pom.xml we therefore should define
it in the plugin management. This makes the spotless plugin available to
all sub-poms including the parent/pom and its children.

By defining the version in the bom/pom.xml we can remove the version
definition from the parent/pom.xml.

Lastly, since the spotless plugin is now managed by the bom/pom.xml,
I've moved further configuration added by the parent/pom.xml into the
build/plugins tag to avoid managing the plugin in 2 different poms.