test(engine): only verify the job canceled event
Since the command is no longer written, the test cases that were
verifying that the command was written should be updated to verify that
the job canceled event is written in the same spot.

Note that in the BoundaryEventTest it highlights the difference between
asynchronously canceling the job with a command and synchronously
(immediately) canceling the job with an event.