fix(engine): also cancel job in error_thrown state
Jobs in the error_thrown state should also be able to be cancelled. By
adding it to this list, we make sure that the job with error_thrown
state can be cancelled by the engine when the element it belongs to is
being terminated.

(cherry picked from commit 57fc9da294dd7c025a599ab5ad54747b6f7e6db0)