fix(engine): cancel job synchronously
We need to make sure the job is cancelled immediately (synchronously)
because otherwise the IncidentResolvedApplier may still make the job
activatable again.

To cancel the job immediately we simply need to use the statewriter to
write Job Cancelled, instead of using the commandwriter to write Cancel
Job. The statewriter will immediately apply the event when writing it.