refactor(engine): prevent instant rescheduling
Before this change, the delay calculated to reschedule a task could be negative or close to 0.
This lead to the checker being immediately rescheduled. This is bad, because it does not leave room
for other tasks to run.

With this change, a lower floor is applied when the task is rescheduled.

(cherry picked from commit 0f4a6ab7d6df052b07d2d22ecde56a650be54753)