refactor(engine): deprecate cancel job command
We no longer write cancel job commands, only the job canceled event.
However, since the cancel command can still exist on log streams, we
should continue to support it.

Therefore, it's now deprecated for removal in a future version. Note
that we should only remove the intent and the processor when we decide
to stop being backwards compatible with existing log streams.