refactor(atomix): remove usage of journal test-jar
Removes usage of journal test-jar via the `LogCorrupter`. While the new
tests is potentially not as accurate, using the journal's log corrupter
utility is essentially leaking implementation details of the journal to
Raft. These tests should not care how the log is corrupted - how to
corrupt it is implementation detail, in a way.

One test was removed - this test was checking if the log would be
truncated on partial writes. Again, this is journal implementation
detail - the journal guarantees that you will never read corrupted
entries successfully, and also that partial writes are correctly
detected and handled. This is all tested in the journal module, more
specifically in `JournalTest#shouldDeletePartiallyWrittenEntry` and
`JournalTest#shouldDetectCorruptedEntry`.

The tests should be refactored long term to be journal implementation
agnostic, but as is, they're fine.

Additionally, we should not use test JARs in general. Not only does that
import test classes, but most of the classes or utilities we write for
one off tests are often not meant to be used by other tests - they're
not documented, poorly tested, and since we don't expect them to be used
else where, we're more likely to change the behavior.

Using test JARs in another module is not only further coupling the
modules, but it's worse as it usually couples implementations together.
Now not only are we coupling to the API of the module, but to its test
utilities and most likely to its implementation details. Most likely
because usually that utility was created to test the implementation of
said module.