build: clean up POMs
Primarily cleans up the parent module POM, removing dependencies that
were added for convergence but are not required anymore. Dependencies
added for convergence are now put in their own section to make this task
easier in the future, along with comments on which libraries had
conflicts.

Additionally, for some POMs which were modified due to the above tasks,
I've separated production and test dependencies. Again this is to make
it easier to perform cleanup and maintenance, but it also helps us more
quickly understand what exactly will be included in the final JAR at a
glance.

Finally, modules which generated SBE code were slightly fixed - we can
look at the Wiki page for SBE Tool with Maven, which proposed a slightly
different approach that doesn't require us to specify sbe-tool as a
dependency. We can tell the exec plugin to use the plugin dependencies
instead of the project's dependencies, and then sbe-tool is just a
dependency of that plugin.