test(engine): should observe job lifetime
While trying to write a test that verifies that a record filter is set
on the metrics exporter, I kinda stumbled into writing a test that
verifies one of the default behaviors of the metrics exporter: it should
be able to observe job lifetime.

I was unable to find other tests for this code, only for the metrics
exporter configuration. So I felt it made sense to add this test case.

(cherry picked from commit c8f1eca0f0623e8538eaf3b4de6c3528fac6d9af)