refactor(broker): remove knowledge of log stream
Moves `ZeebeEntryValidator` to a broker `raft` specific package, and
removes any knowledge of log stream specifics.

The validator now only uses the entry's lowest and highest position
properties to determine if there are gaps between consecutive
`ApplicationEntry`. The check to guarantee there are no gaps in a single
batch will be moved to the `LogStorageAppender` itself.