refactor(protocol-test-util): only generate positive longs
Ensure long-typed properties of generated records are always positive.
This is a simple workaround, as many of the properties of type long are
actually timestamps, or are expected to behave like timestamps. Since
other parts will expect these to be positive, at the very least, we can
do that here.