feat(exporter-test): add exporter API test implementations
Adds test implementation for the exporter API, allowing easier testing
of exporters.

Configuration/context is kept to a minimum: users can set the
configuration that will be instantiated beforehand, but no instantiation
is done for them from the provided args. This is a limitation at the
moment, to avoid coupling the broker's implementation and this one here.
It's sufficient for our own use cases for now.

The controller can schedule tasks and process them deterministically.
This is done manually, by having users "tick" the controller manually -
it will then run all scheduled tasks which have "expired" in order,
synchronously. Tasks are kept in memory, even after being executed, so
we can still assert properties of previous scheduled, canceled, or
executed tasks.