refactor(exporters): extract record index routing logic
Extracts the record index routing logic from the `ElasticsearchClient`
class into a new class.

For now, we use a new router instance in the abstract test case - this
will later be removed as the abstract test case will be completely
removed, so for now we just keep it to to keep refactoring to a minimum
in this commit.