test(logstreams): fix race condition with failure listener
Refactors the test to await the actor close instead of using a failure
listener. Using the failure listener is better/more accurate, however
since we cannot control when the failure listener is added vs when the
subscription is polled, we cannot guarantee that the failure is injected
AFTER the listener has been added.

Instead, we change the test to check that the actor closes within 10
seconds, which is what would happen if an inconsistency is detected.