build: move spotless config and profiles to BOM
Moves the spotless and license plugin configuration to the BOM,
and also the autoFormat and checkFormat profiles. Removes the
`sortPom` module explicit dependency and configuration in favor
of the built in sortPom extension from spotless, which is the
same backend in the end.

This means we also move the autoFormat/checkFormat profiles into
the BOM. Since the autoFormat profile is active by default, the
`skipChecks` property for these three plugins is also defined there
from now on. I've left the `skipChecks` property in `parent` as
well. I know it seems redundant, but I felt otherwise it seemed to
come out of "nowhere", but feel free to challenge that.

The BOM is now sorted using sortPom, and for the rest of the modules
everything else applies and stays the same. There was no need to add
`combine.children` directives anywhere, since what we do want is the
default, i.e. merge, not append.