refactor(exporters): reuse same bulkRequest instance
Reuses the same bulk request instance instead of allocating a new one.
Many of our tests rely on inspecting the bulkRequest list to see what
has been staged there. While this is not ideal, we can leave it as is
for now, but it's important to guarantee that we can properly verify
what's been staged in memory.