refactor(exporters): use DTO for bulk request items
Instead of manually creating bulk request item pairs, use a DTO
specifically made for this, `BulkRequestAction`. This allows us to later
better test that the action is as expected, since we now have a fixed
data format we can use to compare instead of a free-form map.

It also simplifies reading/understanding the original code, and will
ease maintenance in the future should the DTO need to be changed (as
there's a clear place to do so, including controlling how it is
serialized).