fix(engine): use target decision id when drg invalid
When the decision evaluation fails, a decision evaluation FAILED event
is written with information about the failure. This includes the
id of the decision that failed.

However, when the eval fails because the DRG is invalid, then we
would use null as this decisionId, which cannot be written into the
event and leads to an NPE.

If the decisision evaluation fails because the drg is invalid, then we
still know about a decision id that failed. We can simply use the
decision id of the targetted decision.

(cherry picked from commit 937dab2657143a8e2daee52f8f52d96a2a171171)