fix(engine): use target decision id when audit log is empty
When the decision evaluation fails, we write this failure in an event
with relevant information. This includes the id of the decision that
failed. If this value is null, we cannot write the event because of a
NPE.

Instead of null we can use the target decision's id as the default, in
case the evaluation failed for some reason. If we then can figure out
which decision actually failed (perhaps a required decision of the
target decision failed), then we can use that id instead.

(cherry picked from commit f3106c4c139244961d1b341412451779fb6970c9)