feat: return empty topology response when not known
The topology is a way to view the cluster from the gateway's point of view. Returning UNAVAILABLE hides this and makes it hard to distinguish between a call where the gateway is unavailable and where the topology is simply not known. By returning an empty topology we can differentiate between an unavailable gateway and a not known topology.