fix(engine): do no create writer when streamprocessor is in replay mode
In follower role, StreamProcessor is running only in replay mode. When the writer is created, dispatcher is also opened which allocates direct buffer. This is unnecessary as the writer is never used. The allocated buffer consumes memory and can create memory pressure on the system. To fix this, we create the writer only after the replay is completed.

(cherry picked from commit c14e56fdd2ca9a5bb40ded47983c4c55980883a9)