test(engine): use streamWriter factory in StreamProcessorHealthTest
After the fix where the creation of the writer is moved after the replay is completed, this test failed because the mock writer created by the test is overwritten by the StreamProcessor. To prevent this, we added a new streamWriter factor that can be used by the test to inject mock writers.

(cherry picked from commit d85da0293aa37483108e7fe22fc3c9613a821f1a)