test(broker): metrics exporter should config a filter
The metrics exporter should define a filter so that the exporter
director doesn't unnecessarily read records for the metrics exporter.

This test does not yet add any specific recordtype recordvalue
combinations that are asserted and will fail anyways, because the
metrics exporter does not yet define a filter. However, I wanted to
seperate the test setup from the actual filter choices.