test(exporters): refactor integration tests
Refactors the `ElasticsearchExporter` integration tests. Almost all
tests are now done in `ElasticsearchExporterIT`, with a static container
shared by all tests. The flakiness is removed as everything happens
synchronously in the test thread, except for the inherent asynchronocity
with the Elastic instance.

There is one other class, `FaultToleranceIT`, where tests which need to
change the configuration of the Elastic instance or start/stop it are
placed, since these won't play well with the other ones sharing the same
instance.