fix: resolve cherry-pick conflicts
Cherrypicking the exporter API test implementations with 2ea243b1a from
041c5226f suffered from some conflicts. I've extracted the conflict
resolutions to a separate commit to highlight and clarify these.

Specificially, the parent pom had a merge conflict that was easily
resolved. For the cherrypicked code to work we need the jcip dependency.

In addition, the newly added zeebe-exporter-test pom was referring to
the newer parent pom's snapshot version (8.1.0-snapshot), but when
cherrypicked to 8.0.3-snapshot, we should of course use that version.