fix: prevent duplicate events for drgs and decisions
When writing the records after deploying a new DRG we would iterate over all the drgs and decisions in the DeploymentRecord. New drgs and decisions get gradually appended to the DeploymentRecord for each resource that's being processed. Since we iterate over this for every resource, it causes the same event to be send multiple times. Once for each of the drgs deployed.

By filtering the drg metadata that is available in the DeploymentRecord to only include the resource that we are currently transforming we can stop this behavior and support deploying multiple drgs at once.

(cherry picked from commit 223d5c7f828dc30659e46eb47feb7b69408ff147)