ci(.github): add publish-test-results workflow
We want to publish the test results from test runs. The ZPT repo already
has a workflow defined that does this. It runs when the 'Build and test'
workflow has completed and was not skipped. Then it downloads and
extracts the artifacts from the original workflow run and runs the
publish-unit-test-result-action against them.

The workflow is set up like this to support PRs from forked repos as
well as dependabot PRs. A comment in the top of the workflow definition
describes this.

For clarity, this commit is a direct copy of the file from the ZPT repo.
This makes it easier to distinquish the changes we've had to make in the
zeebe repo. This first commit will therefore introduce a workflow that
doesn't yet run because no workflow named 'Build and test' exists in the
zeebe repo.