ci(.github): always publish test results
We want to show the test results with the new
publish-unit-test-result-action when the tests failed, but also when all
tests passed. This might provide additional crucial information about
the CI, e.g. if no tests were run then the CI could've passed, this
report can highlight this.

To make that possible, we need to publish the test results in all
cases. We can do that using `if: always` which guarantees that the step
is executed both when the step failed and when it was successful.