test: append entries of predictable size
Serializing a randomly generated string to bytes produces byte arrays
of unpredictable size. This does not match the expectation of
`appendEntry` which takes an `entrySize` parameter.

With this commit we now directly generate a random byte array of the
exact size given in `entrySize`.

In two tests, the number of appended entries has to increase to reliably
compact the log on snapshot

(cherry picked from commit 1f33a9db66f271732ecbb37018b55fd6c251b330)