test: let embedded broker rule wait for healthy leader
Previously, the rule only waited for a leader of partition 1 before
returning from `startBroker`. With this change, the leader must also
report as healthy.

This is intended to fix some flaky tests where after the startup,
commands send to the leader of partition 1 failed because no command-api
handler was registered. Since registration of handlers is required for
a healthy status, this should hopefully prevent these kinds of flaky
tests.