fix: wait for partition listeners before completing recovery
Previously, partition listeners were notified on transitions but the
`onRecoveredInternal` was called without waiting for the result.
Since we use partition listeners for installing, among others, the
command api services, it is important to wait on that to complete
before finishing the recovery and marking services as installed.

This ensures that the partition is only marked as healthy when the
command api services are registered.