test(qa): fix IPv6 flakiness
The IPv6 test is sometimes flaky because it may take a while for a full
cluster to start: 3 brokers and one gateway. To test IPv6, however, we
don't really need so many nodes, and having 1 broker and 1 gateway is
enough: all ports are tested properly (except the management port, which
is tested already by Spring).

Additionally, this fixes that if the test fails once, the network may
not be removed properly afterwards, which causes further iterations to
fail. We ensure the network is closed here, and also we print out a
proper error if we fail to start the network on the next run.

Finally I took the time to clean up the test using new APIs provided by
zeebe-test-container.