feat(gateway): map start instructions for CreateProcessInstanceWithResult
Initially, the plan was to omit this mapping to reduce scope. However, when working on #9398 we
decided to include the mapping. Reason is that both the request to create a process instance and
the request to create a process instance with result share code, which enables both to take
start instructions. On balance, we prefer to support it for both cases rather than implementing
additional logic to remove this capability from one of the cases.