test(atomix): fix flaky commit notify test
Use commitIndex to verify that listener is notified instead of using called count. It is possible that one of the follower committed the InitialEntry after the CommitListener is added. As a result the test failed because the expected call count is different.

(cherry picked from commit 6f3f662b73ee683abe42cca8a08384840dfa1cdf)