test: do not restart broker on snapshot test
Previously we took no snapshot if the exporter hasn't commited the positions, but this is no longer true.
This means if we run the DebugExporter we still can take snapshots.

This allows to reduce the complexity in the SnapshotTest, as it is no longer necessary to restart the Broker
without the DebugExporter. This will of course also reduce the execution time of the test, and hopefully the flakiness.