ci(.github): run CodeQL job on selfhosted runner
The CodeQL job is rather slow (sometimes exceeding 20 min for just the
Build step. To speed up things we should run it on our selfhosted gha
runners. Let's start with the smallest runner first (8cpu).

We can also use the Nexus cache for this job to further speed it up and
make sure the artifacts are taken from our own nexus.

Lastly, this requires us to set up maven explicitly as it is installed
by default on the GH runners, but not on our selfhosted runners.